requirements.txt
pymongo==4.13.2
fastapi
uvicorn
pydantic
requests
numpy
python-dotenv
openai
google-genai>=1.0.0
pandas
ipython
pyyaml
pyprojroot
langsmith
langchain-google-genai
motor
scipy
scikit-learn
langchain